feat: extract PhosphorRendering library from PlasmaZones rendering pipeline
Extract the GPU shader rendering pipeline into a standalone LGPL-2.1
PhosphorRendering library, completing the three-library architecture:

  PhosphorShell (layer-shell, registry, BaseUniforms)
    └─ PhosphorRendering (RHI pipeline, multipass, textures, compiler)
         └─ PlasmaZones (zones, overlays, tiling, D-Bus)

## PhosphorRendering library (3,733 LOC, LGPL-2.1)

- ShaderEffect: QQuickItem with 35 Shadertoy-compatible Q_PROPERTYs,
  QML_ELEMENT, uniform extension support, custom params/colors
- ShaderNodeRhi: QSGRenderNode with full RHI rendering — VBO, UBO,
  SRBs, pipelines, multipass (single/ping-pong/multi-buffer), texture
  management (audio, user, wallpaper, depth), partial UBO uploads
- ShaderCompiler: static GLSL→SPIR-V compiler with thread-safe cache,
  include expansion, SPIR-V 1.0 + GLSL 330 + ES 300/310/320 targets
- Indexed setter API: setCustomParams(int, QVector4D) replaces 40 named
  virtual methods; setExtraBinding() for consumer texture slots

## PlasmaZones migration (4,344 LOC deleted, 466 LOC added)

- ZoneShaderItem now inherits PhosphorRendering::ShaderEffect
- ZoneShaderNodeRhi now inherits PhosphorRendering::ShaderNodeRhi
- ZoneUniformExtension implements IUniformExtension for zone UBO data
  (4096 bytes: zoneRects, zoneFillColors, zoneBorderColors, zoneParams)
- Labels texture at binding 1 via setExtraBinding(1, tex, sampler)
- appField0/appField1 for zoneCount/highlightedCount
- Deleted: zoneshadernodebase.h, rendering_macros.h, and 4 split TUs
  (pipeline.cpp, uniforms.cpp, setters.cpp, internal.h)
- Removed PhosphorShell ShaderEffect/RenderNode stubs (now in PR lib)

## Packaging

- RPM: libPhosphorRendering.so, headers, dev files
- Debian: phosphorrendering/* → LGPL-2.1-or-later
- Nix: license list updated

93/93 tests pass, zero compiler warnings.
