+++ packaging/debian/copyright
@@ -11,6 +11,10 @@ Files: phosphorshell/*
 Copyright: 2026 fuddlesworth
 License: LGPL-2.1-or-later
 
+Files: phosphorrendering/*
+Copyright: 2026 fuddlesworth
+License: LGPL-2.1-or-later
+
 Files: debian/*
 Copyright: 2026 fuddlesworth
 License: GPL-3.0-or-later
