+++ phosphorrendering/include/PhosphorRendering/ShaderCompiler
@@ -1,0 +1,4 @@
+// SPDX-FileCopyrightText: 2026 fuddlesworth
+// SPDX-License-Identifier: LGPL-2.1-or-later
+#pragma once
+#include <PhosphorRendering/ShaderCompiler.h>
