+++ phosphorrendering/include/PhosphorRendering/ShaderNodeRhi
@@ -1,4 +1,4 @@
 // SPDX-FileCopyrightText: 2026 fuddlesworth
 // SPDX-License-Identifier: LGPL-2.1-or-later
 #pragma once
-#include <PhosphorShell/ShaderEffect.h>
+#include <PhosphorRendering/ShaderNodeRhi.h>
