+++ phosphorrendering/src/shadernoderhisetters.cpp
@@ -1,105 +1,62 @@
 // SPDX-FileCopyrightText: 2026 fuddlesworth
-// SPDX-License-Identifier: GPL-3.0-or-later
+// SPDX-License-Identifier: LGPL-2.1-or-later
 
-#include "../zoneshadernoderhi.h"
 #include "internal.h"
-#include "../rendering_macros.h"
 
-#include "../../../core/logging.h"
-#include "../../../core/shaderutils.h"
+#include <PhosphorRendering/ShaderCompiler.h>
 
 #include <QFileInfo>
 #include <cmath>
 
-namespace PlasmaZones {
+namespace PhosphorRendering {
 
 // ============================================================================
-// Zone Data Setters
+// Uniform Extension
 // ============================================================================
 
-void ZoneShaderNodeRhi::setZones(const QVector<ZoneData>& zones)
+void ShaderNodeRhi::setUniformExtension(std::shared_ptr<PhosphorShell::IUniformExtension> extension)
 {
-    int count = qMin(zones.size(), MaxZones);
-    m_zones = zones.mid(0, count);
-    m_uniformsDirty = true;
-    m_zoneDataDirty = true;
-}
-
-void ZoneShaderNodeRhi::setZone(int index, const ZoneData& data)
-{
-    if (index >= 0 && index < MaxZones) {
-        if (index >= m_zones.size()) {
-            m_zones.resize(index + 1);
-        }
-        m_zones[index] = data;
-        m_uniformsDirty = true;
-        m_zoneDataDirty = true;
-    }
-}
-
-void ZoneShaderNodeRhi::setZoneCount(int count)
-{
-    if (count >= 0 && count <= MaxZones) {
-        m_zones.resize(count);
-        m_uniformsDirty = true;
-        m_zoneDataDirty = true;
-    }
-}
-
-void ZoneShaderNodeRhi::setHighlightedZones(const QVector<int>& indices)
-{
-    m_highlightedIndices = indices;
-    for (int i = 0; i < m_zones.size(); ++i) {
-        m_zones[i].isHighlighted = indices.contains(i);
-    }
-    m_uniformsDirty = true;
-    m_zoneDataDirty = true;
-}
-
-void ZoneShaderNodeRhi::clearHighlights()
-{
-    m_highlightedIndices.clear();
-    for (auto& zone : m_zones) {
-        zone.isHighlighted = false;
-    }
-    m_uniformsDirty = true;
-    m_zoneDataDirty = true;
+    m_uniformExtension = std::move(extension);
+    // Force UBO recreation with new size on next prepare()
+    m_ubo.reset();
+    m_initialized = false;
+    m_didFullUploadOnce = false;
+    resetAllBindingsAndPipelines();
 }
 
 // ============================================================================
 // Timing / Resolution / Mouse Setters
 // ============================================================================
 
-void ZoneShaderNodeRhi::setTime(double time)
+void ShaderNodeRhi::setTime(double time)
 {
     m_time = time;
     m_uniformsDirty = true;
     m_timeDirty = true;
-    // Detect wrap-offset change (once per kShaderTimeWrap seconds). When it
-    // advances, iTimeHi gets its own partial upload — see K_TIME_HI_OFFSET.
-    const float newTimeHi = static_cast<float>(std::floor(time / kShaderTimeWrap) * kShaderTimeWrap);
+    const float newTimeHi =
+        static_cast<float>(std::floor(time / PhosphorShell::kShaderTimeWrap) * PhosphorShell::kShaderTimeWrap);
     if (newTimeHi != m_timeHi) {
         m_timeHi = newTimeHi;
         m_timeHiDirty = true;
     }
-    // iDate is in the scene data region but is recomputed from the system clock
-    // in syncUniformsFromData(), so it updates whenever m_sceneDataDirty is set.
-    // Do NOT set m_zoneDataDirty — the time block has its own upload region.
     m_sceneDataDirty = true;
 }
-void ZoneShaderNodeRhi::setTimeDelta(float delta)
+
+void ShaderNodeRhi::setTimeDelta(float delta)
 {
     m_timeDelta = delta;
     m_uniformsDirty = true;
     m_timeDirty = true;
 }
-void ZoneShaderNodeRhi::setFrame(int frame)
+
+void ShaderNodeRhi::setFrame(int frame)
 {
     m_frame = frame;
     m_uniformsDirty = true;
     m_timeDirty = true;
 }
-void ZoneShaderNodeRhi::setResolution(float width, float height)
+
+void ShaderNodeRhi::setResolution(float width, float height)
 {
     if (m_width != width || m_height != height) {
         m_width = width;
@@ -108,7 +65,8 @@ void ZoneShaderNodeRhi::setResolution(float width, float height)
         m_sceneDataDirty = true;
     }
 }
-void ZoneShaderNodeRhi::setMousePosition(const QPointF& pos)
+
+void ShaderNodeRhi::setMousePosition(const QPointF& pos)
 {
     m_mousePosition = pos;
     m_uniformsDirty = true;
@@ -116,51 +74,89 @@ void ZoneShaderNodeRhi::setMousePosition(const QPointF& pos)
 }
 
 // ============================================================================
-// Custom Params (DRY macro — 4 invocations replace 24 lines)
+// Custom Params (indexed API)
 // ============================================================================
 
-RHINODE_PARAMS_SETTER(CustomParams1, m_customParams1)
-RHINODE_PARAMS_SETTER(CustomParams2, m_customParams2)
-RHINODE_PARAMS_SETTER(CustomParams3, m_customParams3)
-RHINODE_PARAMS_SETTER(CustomParams4, m_customParams4)
-RHINODE_PARAMS_SETTER(CustomParams5, m_customParams5)
-RHINODE_PARAMS_SETTER(CustomParams6, m_customParams6)
-RHINODE_PARAMS_SETTER(CustomParams7, m_customParams7)
-RHINODE_PARAMS_SETTER(CustomParams8, m_customParams8)
-
-// ============================================================================
-// Custom Colors (DRY macro — 8 invocations replace 48 lines)
-// ============================================================================
-
-RHINODE_COLOR_SETTER(CustomColor1, m_customColor1)
-RHINODE_COLOR_SETTER(CustomColor2, m_customColor2)
-RHINODE_COLOR_SETTER(CustomColor3, m_customColor3)
-RHINODE_COLOR_SETTER(CustomColor4, m_customColor4)
-RHINODE_COLOR_SETTER(CustomColor5, m_customColor5)
-RHINODE_COLOR_SETTER(CustomColor6, m_customColor6)
-RHINODE_COLOR_SETTER(CustomColor7, m_customColor7)
-RHINODE_COLOR_SETTER(CustomColor8, m_customColor8)
-RHINODE_COLOR_SETTER(CustomColor9, m_customColor9)
-RHINODE_COLOR_SETTER(CustomColor10, m_customColor10)
-RHINODE_COLOR_SETTER(CustomColor11, m_customColor11)
-RHINODE_COLOR_SETTER(CustomColor12, m_customColor12)
-RHINODE_COLOR_SETTER(CustomColor13, m_customColor13)
-RHINODE_COLOR_SETTER(CustomColor14, m_customColor14)
-RHINODE_COLOR_SETTER(CustomColor15, m_customColor15)
-RHINODE_COLOR_SETTER(CustomColor16, m_customColor16)
-
-// ============================================================================
-// Labels / Audio / User Texture Setters
-// ============================================================================
-
-void ZoneShaderNodeRhi::setLabelsTexture(const QImage& image)
+void ShaderNodeRhi::setCustomParams(int index, const QVector4D& params)
 {
-    m_labelsImage = image;
-    m_labelsTextureDirty = true;
+    if (index < 0 || index >= kMaxCustomParams) {
+        return;
+    }
+    if (m_customParams[index] == params) {
+        return;
+    }
+    m_customParams[index] = params;
     m_uniformsDirty = true;
+    m_sceneDataDirty = true;
 }
 
-void ZoneShaderNodeRhi::setAudioSpectrum(const QVector<float>& spectrum)
+// ============================================================================
+// Custom Colors (indexed API)
+// ============================================================================
+
+void ShaderNodeRhi::setCustomColor(int index, const QColor& color)
+{
+    if (index < 0 || index >= kMaxCustomColors) {
+        return;
+    }
+    if (m_customColors[index] == color) {
+        return;
+    }
+    m_customColors[index] = color;
+    m_uniformsDirty = true;
+    m_sceneDataDirty = true;
+}
+
+// ============================================================================
+// App Fields
+// ============================================================================
+
+void ShaderNodeRhi::setAppField0(int value)
+{
+    if (m_baseUniforms.appField0 == value) {
+        return;
+    }
+    m_baseUniforms.appField0 = value;
+    m_uniformsDirty = true;
+    m_sceneDataDirty = true;
+}
+
+void ShaderNodeRhi::setAppField1(int value)
+{
+    if (m_baseUniforms.appField1 == value) {
+        return;
+    }
+    m_baseUniforms.appField1 = value;
+    m_uniformsDirty = true;
+    m_sceneDataDirty = true;
+}
+
+// ============================================================================
+// Extra Bindings (consumer-managed)
+// ============================================================================
+
+void ShaderNodeRhi::setExtraBinding(int binding, QRhiTexture* texture, QRhiSampler* sampler)
+{
+    m_extraBindings[binding] = ExtraBinding{texture, sampler};
+    m_extraBindingsDirty = true;
+    resetAllBindingsAndPipelines();
+}
+
+void ShaderNodeRhi::removeExtraBinding(int binding)
+{
+    auto it = m_extraBindings.find(binding);
+    if (it != m_extraBindings.end()) {
+        m_extraBindings.erase(it);
+        m_extraBindingsDirty = true;
+        resetAllBindingsAndPipelines();
+    }
+}
+
+// ============================================================================
+// Audio / User Texture / Wallpaper Setters
+// ============================================================================
+
+void ShaderNodeRhi::setAudioSpectrum(const QVector<float>& spectrum)
 {
     if (m_audioSpectrum == spectrum) {
         return;
@@ -170,23 +166,21 @@ void ZoneShaderNodeRhi::setAudioSpectrum(const QVector<float>& spectrum)
     m_uniformsDirty = true;
 }
 
-void ZoneShaderNodeRhi::setUserTexture(int slot, const QImage& image)
+void ShaderNodeRhi::setUserTexture(int slot, const QImage& image)
 {
     if (slot < 0 || slot >= kMaxUserTextures) {
         return;
     }
-    // QImage uses implicit sharing; cacheKey() is O(1) and avoids
-    // re-uploading the identical texture to the GPU every frame.
     if (m_userTextureImages[slot].cacheKey() == image.cacheKey()) {
         return;
     }
     m_userTextureImages[slot] = image;
     m_userTextureDirty[slot] = true;
     m_uniformsDirty = true;
-    m_zoneDataDirty = true;
+    m_sceneDataDirty = true;
 }
 
-void ZoneShaderNodeRhi::setUserTextureWrap(int slot, const QString& wrap)
+void ShaderNodeRhi::setUserTextureWrap(int slot, const QString& wrap)
 {
     if (slot < 0 || slot >= kMaxUserTextures) {
         return;
@@ -196,17 +190,12 @@ void ZoneShaderNodeRhi::setUserTextureWrap(int slot, const QString& wrap)
         return;
     }
     m_userTextureWraps[slot] = use;
-    // Force sampler recreation with new wrap mode
     m_userTextureSamplers[slot].reset();
     resetAllBindingsAndPipelines();
 }
 
-void ZoneShaderNodeRhi::setWallpaperTexture(const QImage& image)
+void ShaderNodeRhi::setWallpaperTexture(const QImage& image)
 {
-    // Compare by data pointer and size: convertToFormat() creates a new QImage
-    // with a different cacheKey() even for identical content, so cacheKey() is
-    // unreliable. QImage uses implicit sharing, so constBits() identity + size
-    // is a fast check for the same underlying data.
     if (m_wallpaperImage.constBits() == image.constBits() && m_wallpaperImage.size() == image.size()) {
         return;
     }
@@ -215,92 +204,38 @@ void ZoneShaderNodeRhi::setWallpaperTexture(const QImage& image)
     m_uniformsDirty = true;
 }
 
-void ZoneShaderNodeRhi::setUseWallpaper(bool use)
+void ShaderNodeRhi::setUseWallpaper(bool use)
 {
     if (m_useWallpaper == use) {
         return;
     }
     m_useWallpaper = use;
-    // Toggling wallpaper adds/removes binding 11 from the SRB layout.
-    // resetAllBindingsAndPipelines() resets all SRBs AND pipelines so they are
-    // recreated with the new layout.
     resetAllBindingsAndPipelines();
     markDirty(QSGNode::DirtyMaterial);
 }
 
-void ZoneShaderNodeRhi::appendUserTextureBindings(QVector<QRhiShaderResourceBinding>& bindings) const
-{
-    for (int t = 0; t < kMaxUserTextures; ++t) {
-        if (m_userTextures[t] && m_userTextureSamplers[t]) {
-            bindings.append(QRhiShaderResourceBinding::sampledTexture(7 + t, QRhiShaderResourceBinding::FragmentStage,
-                                                                      m_userTextures[t].get(),
-                                                                      m_userTextureSamplers[t].get()));
-        }
-    }
-}
-
-void ZoneShaderNodeRhi::appendWallpaperBinding(QVector<QRhiShaderResourceBinding>& bindings) const
-{
-    if (m_useWallpaper && m_wallpaperTexture && m_wallpaperSampler) {
-        bindings.append(QRhiShaderResourceBinding::sampledTexture(11, QRhiShaderResourceBinding::FragmentStage,
-                                                                  m_wallpaperTexture.get(), m_wallpaperSampler.get()));
-    }
-}
-
-void ZoneShaderNodeRhi::appendDepthBinding(QVector<QRhiShaderResourceBinding>& bindings) const
-{
-    if (m_useDepthBuffer && m_depthTexture && m_depthSampler) {
-        bindings.append(QRhiShaderResourceBinding::sampledTexture(12, QRhiShaderResourceBinding::FragmentStage,
-                                                                  m_depthTexture.get(), m_depthSampler.get()));
-    }
-}
-
-void ZoneShaderNodeRhi::setUseDepthBuffer(bool use)
+void ShaderNodeRhi::setUseDepthBuffer(bool use)
 {
     if (m_useDepthBuffer == use) {
         return;
     }
     m_useDepthBuffer = use;
-    // Toggling depth buffer adds/removes binding 12 from the SRB layout.
-    // Release depth resources so ensureBufferTarget() recreates them with the new config.
     m_depthTexture.reset();
     m_depthSampler.reset();
-    // resetAllBindingsAndPipelines() resets all SRBs AND pipelines so they are
-    // recreated with the new layout.
     resetAllBindingsAndPipelines();
     markDirty(QSGNode::DirtyMaterial);
 }
 
-void ZoneShaderNodeRhi::resetAllBindingsAndPipelines()
-{
-    m_srb.reset();
-    m_srbB.reset();
-    m_bufferSrb.reset();
-    m_bufferSrbB.reset();
-    // Pipelines bake the SRB layout (descriptor set layout) at creation time.
-    // When a texture binding is added or removed (e.g. labels texture created
-    // late, audio spectrum resized from zero), the rebuilt SRBs have a different
-    // layout than the pipelines expect. Vulkan rejects the mismatch silently,
-    // causing missing/incomplete rendering. Reset all pipelines so they are
-    // recreated with the new SRB layout.
-    m_pipeline.reset();
-    m_bufferPipeline.reset();
-    for (int i = 0; i < kMaxBufferPasses; ++i) {
-        m_multiBufferSrbs[i].reset();
-        m_multiBufferPipelines[i].reset();
-    }
-}
-
 // ============================================================================
-// Buffer Shader Path / Feedback / Scale / Wrap
+// Buffer Shader Path / Feedback / Scale / Wrap / Filter
 // ============================================================================
 
-void ZoneShaderNodeRhi::setBufferShaderPath(const QString& path)
+void ShaderNodeRhi::setBufferShaderPath(const QString& path)
 {
     setBufferShaderPaths(path.isEmpty() ? QStringList() : QStringList{path});
 }
 
-void ZoneShaderNodeRhi::setBufferShaderPaths(const QStringList& paths)
+void ShaderNodeRhi::setBufferShaderPaths(const QStringList& paths)
 {
     QStringList trimmed;
     for (int i = 0; i < qMin(paths.size(), kMaxBufferPasses); ++i) {
@@ -314,8 +249,8 @@ void ZoneShaderNodeRhi::setBufferShaderPaths(const QStringList& paths)
     m_bufferPaths = trimmed;
     m_bufferPath = trimmed.isEmpty() ? QString() : trimmed.constFirst();
 
-    qCDebug(lcOverlay) << "ZoneShaderNodeRhi setBufferShaderPaths count=" << m_bufferPaths.size()
-                       << "multiBufferMode=" << (m_bufferPaths.size() > 1);
+    qCDebug(lcShaderNode) << "ShaderNodeRhi setBufferShaderPaths count=" << m_bufferPaths.size()
+                          << "multiBufferMode=" << (m_bufferPaths.size() > 1);
 
     m_bufferShaderDirty = true;
     m_bufferShaderReady = false;
@@ -334,7 +269,7 @@ void ZoneShaderNodeRhi::setBufferShaderPaths(const QStringList& paths)
         const QString& path = m_bufferPaths.constFirst();
         if (QFileInfo::exists(path)) {
             QString err;
-            m_bufferFragmentShaderSource = detail::loadAndExpandShader(path, &err);
+            m_bufferFragmentShaderSource = loadAndExpandShader(path, &err);
             if (!m_bufferFragmentShaderSource.isEmpty()) {
                 m_bufferMtime = QFileInfo(path).lastModified().toMSecsSinceEpoch();
             }
@@ -364,7 +299,7 @@ void ZoneShaderNodeRhi::setBufferShaderPaths(const QStringList& paths)
     m_srbB.reset();
 }
 
-void ZoneShaderNodeRhi::setBufferFeedback(bool enable)
+void ShaderNodeRhi::setBufferFeedback(bool enable)
 {
     if (m_bufferFeedback == enable) {
         return;
@@ -381,7 +316,7 @@ void ZoneShaderNodeRhi::setBufferFeedback(bool enable)
     m_bufferFeedbackCleared = false;
 }
 
-void ZoneShaderNodeRhi::setBufferScale(qreal scale)
+void ShaderNodeRhi::setBufferScale(qreal scale)
 {
     const qreal clamped = qBound(0.125, scale, 1.0);
     if (qFuzzyCompare(m_bufferScale, clamped)) {
@@ -409,7 +344,7 @@ void ZoneShaderNodeRhi::setBufferScale(qreal scale)
     }
 }
 
-void ZoneShaderNodeRhi::setBufferWrap(const QString& wrap)
+void ShaderNodeRhi::setBufferWrap(const QString& wrap)
 {
     const QString use = normalizeWrapMode(wrap);
     if (m_bufferWrapDefault == use) {
@@ -423,7 +358,7 @@ void ZoneShaderNodeRhi::setBufferWrap(const QString& wrap)
     resetAllBindingsAndPipelines();
 }
 
-void ZoneShaderNodeRhi::setBufferWraps(const QStringList& wraps)
+void ShaderNodeRhi::setBufferWraps(const QStringList& wraps)
 {
     bool changed = false;
     for (int i = 0; i < kMaxBufferPasses; ++i) {
@@ -439,7 +374,7 @@ void ZoneShaderNodeRhi::setBufferWraps(const QStringList& wraps)
     }
 }
 
-void ZoneShaderNodeRhi::setBufferFilter(const QString& filter)
+void ShaderNodeRhi::setBufferFilter(const QString& filter)
 {
     const QString use = normalizeFilterMode(filter);
     if (m_bufferFilterDefault == use) {
@@ -453,7 +388,7 @@ void ZoneShaderNodeRhi::setBufferFilter(const QString& filter)
     resetAllBindingsAndPipelines();
 }
 
-void ZoneShaderNodeRhi::setBufferFilters(const QStringList& filters)
+void ShaderNodeRhi::setBufferFilters(const QStringList& filters)
 {
     bool changed = false;
     for (int i = 0; i < kMaxBufferPasses; ++i) {
@@ -473,10 +408,10 @@ void ZoneShaderNodeRhi::setBufferFilters(const QStringList& filters)
 // Shader Loading (Vertex / Fragment)
 // ============================================================================
 
-bool ZoneShaderNodeRhi::loadVertexShader(const QString& path)
+bool ShaderNodeRhi::loadVertexShader(const QString& path)
 {
     QString err;
-    m_vertexShaderSource = detail::loadAndExpandShader(path, &err);
+    m_vertexShaderSource = loadAndExpandShader(path, &err);
     if (m_vertexShaderSource.isEmpty()) {
         m_shaderError = err.startsWith(QStringLiteral("Failed to open:"))
             ? QString(QStringLiteral("Failed to open vertex shader: ") + path)
@@ -489,10 +424,10 @@ bool ZoneShaderNodeRhi::loadVertexShader(const QString& path)
     return true;
 }
 
-bool ZoneShaderNodeRhi::loadFragmentShader(const QString& path)
+bool ShaderNodeRhi::loadFragmentShader(const QString& path)
 {
     QString err;
-    m_fragmentShaderSource = detail::loadAndExpandShader(path, &err);
+    m_fragmentShaderSource = loadAndExpandShader(path, &err);
     if (m_fragmentShaderSource.isEmpty()) {
         m_shaderError = err.startsWith(QStringLiteral("Failed to open:"))
             ? QString(QStringLiteral("Failed to open fragment shader: ") + path)
@@ -505,7 +440,7 @@ bool ZoneShaderNodeRhi::loadFragmentShader(const QString& path)
     return true;
 }
 
-void ZoneShaderNodeRhi::setVertexShaderSource(const QString& source)
+void ShaderNodeRhi::setVertexShaderSource(const QString& source)
 {
     if (m_vertexShaderSource != source) {
         m_vertexShaderSource = source;
@@ -517,7 +452,7 @@ void ZoneShaderNodeRhi::setVertexShaderSource(const QString& source)
     }
 }
 
-void ZoneShaderNodeRhi::setFragmentShaderSource(const QString& source)
+void ShaderNodeRhi::setFragmentShaderSource(const QString& source)
 {
     if (m_fragmentShaderSource != source) {
         m_fragmentShaderSource = source;
@@ -529,28 +464,55 @@ void ZoneShaderNodeRhi::setFragmentShaderSource(const QString& source)
     }
 }
 
-bool ZoneShaderNodeRhi::isShaderReady() const
+bool ShaderNodeRhi::isShaderReady() const
 {
     return m_shaderReady;
 }
 
-QString ZoneShaderNodeRhi::shaderError() const
+QString ShaderNodeRhi::shaderError() const
 {
     return m_shaderError;
 }
 
-void ZoneShaderNodeRhi::invalidateShader()
+void ShaderNodeRhi::invalidateShader()
 {
     m_shaderDirty = true;
 }
 
-void ZoneShaderNodeRhi::invalidateUniforms()
+void ShaderNodeRhi::invalidateUniforms()
 {
     m_uniformsDirty = true;
     m_timeDirty = true;
     m_timeHiDirty = true;
-    m_zoneDataDirty = true;
+    m_extensionDirty = true;
     m_sceneDataDirty = true;
 }
 
-} // namespace PlasmaZones
+// ============================================================================
+// Shader Include Paths
+// ============================================================================
+
+void ShaderNodeRhi::setShaderIncludePaths(const QStringList& paths)
+{
+    m_shaderIncludePaths = paths;
+}
+
+// ============================================================================
+// Normalize helpers (static)
+// ============================================================================
+
+QString ShaderNodeRhi::normalizeWrapMode(const QString& wrap)
+{
+    return (wrap == QLatin1String("repeat")) ? QStringLiteral("repeat") : QStringLiteral("clamp");
+}
+
+QString ShaderNodeRhi::normalizeFilterMode(const QString& filter)
+{
+    if (filter == QLatin1String("nearest"))
+        return QStringLiteral("nearest");
+    if (filter == QLatin1String("mipmap"))
+        return QStringLiteral("mipmap");
+    return QStringLiteral("linear");
+}
+
+} // namespace PhosphorRendering
