--- phosphorshell/src/shadereffect.cpp
@@ -1,7 +1,0 @@
-// SPDX-FileCopyrightText: 2026 fuddlesworth
-// SPDX-License-Identifier: LGPL-2.1-or-later
-
-// Stub — will be populated by generalizing PlasmaZones ZoneShaderItem
-// during Phase 2 migration.
-
-#include <PhosphorShell/ShaderEffect.h>
