--- src/daemon/rendering/zoneshadernodebase.h
@@ -1,164 +1,0 @@
-// SPDX-FileCopyrightText: 2026 fuddlesworth
-// SPDX-License-Identifier: GPL-3.0-or-later
-
-#pragma once
-
-#include "zoneshadercommon.h"
-#include <QSGRenderNode>
-#include <QColor>
-#include <QImage>
-#include <QString>
-#include <QStringList>
-#include <QVector>
-#include <QVector4D>
-
-namespace PlasmaZones {
-
-/**
- * @brief Abstract base for zone shader render nodes (RHI backend)
- *
- * ZoneShaderItem uses this interface to drive ZoneShaderNodeRhi (QRhi/OpenGL).
- */
-class ZoneShaderNodeBase : public QSGRenderNode
-{
-public:
-    explicit ZoneShaderNodeBase() = default;
-    ~ZoneShaderNodeBase() override = default;
-
-    // Zone data
-    virtual void setZones(const QVector<ZoneData>& zones) = 0;
-    virtual void setZone(int index, const ZoneData& data) = 0;
-    virtual void setZoneCount(int count) = 0;
-    virtual void setHighlightedZones(const QVector<int>& indices) = 0;
-    virtual void clearHighlights() = 0;
-
-    // Timing. setTime takes a double because float32 precision on a seconds-
-    // since-start counter degrades past ~36 h uptime — the node splits it into
-    // iTime (wrapped) + iTimeHi (wrap offset) internally. See kShaderTimeWrap.
-    virtual void setTime(double time) = 0;
-    virtual void setTimeDelta(float delta) = 0;
-    virtual void setFrame(int frame) = 0;
-    virtual void setResolution(float width, float height) = 0;
-    virtual void setMousePosition(const QPointF& pos) = 0;
-
-    // Custom params and colors
-    virtual void setCustomParams1(const QVector4D& params) = 0;
-    virtual void setCustomParams2(const QVector4D& params) = 0;
-    virtual void setCustomParams3(const QVector4D& params) = 0;
-    virtual void setCustomParams4(const QVector4D& params) = 0;
-    virtual void setCustomParams5(const QVector4D& params) = 0;
-    virtual void setCustomParams6(const QVector4D& params) = 0;
-    virtual void setCustomParams7(const QVector4D& params) = 0;
-    virtual void setCustomParams8(const QVector4D& params) = 0;
-    virtual void setCustomColor1(const QColor& color) = 0;
-    virtual void setCustomColor2(const QColor& color) = 0;
-    virtual void setCustomColor3(const QColor& color) = 0;
-    virtual void setCustomColor4(const QColor& color) = 0;
-    virtual void setCustomColor5(const QColor& color) = 0;
-    virtual void setCustomColor6(const QColor& color) = 0;
-    virtual void setCustomColor7(const QColor& color) = 0;
-    virtual void setCustomColor8(const QColor& color) = 0;
-    virtual void setCustomColor9(const QColor& color) = 0;
-    virtual void setCustomColor10(const QColor& color) = 0;
-    virtual void setCustomColor11(const QColor& color) = 0;
-    virtual void setCustomColor12(const QColor& color) = 0;
-    virtual void setCustomColor13(const QColor& color) = 0;
-    virtual void setCustomColor14(const QColor& color) = 0;
-    virtual void setCustomColor15(const QColor& color) = 0;
-    virtual void setCustomColor16(const QColor& color) = 0;
-
-    /** Labels texture (pre-rendered zone numbers). Default no-op for backends that don't support it. */
-    virtual void setLabelsTexture(const QImage& image)
-    {
-        Q_UNUSED(image)
-    }
-
-    /** Audio spectrum (CAVA bar values 0-1). Empty = disabled. Default no-op. */
-    virtual void setAudioSpectrum(const QVector<float>& spectrum)
-    {
-        Q_UNUSED(spectrum)
-    }
-
-    /** User-supplied image texture (bindings 7-10). Slot 0-3. Default no-op. */
-    virtual void setUserTexture(int slot, const QImage& image)
-    {
-        Q_UNUSED(slot)
-        Q_UNUSED(image)
-    }
-    /** Wrap mode for user texture slot: "repeat" or "clamp". Default no-op. */
-    virtual void setUserTextureWrap(int slot, const QString& wrap)
-    {
-        Q_UNUSED(slot)
-        Q_UNUSED(wrap)
-    }
-
-    /** Desktop wallpaper image (binding 11). Only used when shader subscribes via "wallpaper": true. */
-    virtual void setWallpaperTexture(const QImage& image)
-    {
-        Q_UNUSED(image)
-    }
-    /** Whether the shader subscribes to wallpaper texture. */
-    virtual void setUseWallpaper(bool use)
-    {
-        Q_UNUSED(use)
-    }
-
-    /** Enable depth buffer (R32F at binding 12) for MRT depth writing/reading. Default no-op. */
-    virtual void setUseDepthBuffer(bool)
-    {
-    }
-
-    /** Multi-pass: optional buffer pass fragment shader path. No-op if backend does not support multipass. */
-    virtual void setBufferShaderPath(const QString& path)
-    {
-        Q_UNUSED(path)
-    }
-    /** Multi-pass: up to 4 buffer pass fragment shader paths (A→B→C→D). Overrides single path when non-empty. */
-    virtual void setBufferShaderPaths(const QStringList& paths)
-    {
-        Q_UNUSED(paths)
-    }
-    /** When true, buffer pass uses ping-pong (two textures, samples previous frame as iChannel0). Default no-op. */
-    virtual void setBufferFeedback(bool enable)
-    {
-        Q_UNUSED(enable)
-    }
-    /** Buffer resolution scale (e.g. 0.5 = half size). Default no-op. */
-    virtual void setBufferScale(qreal scale)
-    {
-        Q_UNUSED(scale)
-    }
-    /** Buffer channel wrap: "clamp" or "repeat". Default no-op. */
-    virtual void setBufferWrap(const QString& wrap)
-    {
-        Q_UNUSED(wrap)
-    }
-    /** Per-channel buffer wraps (up to 4). Empty = all use setBufferWrap value. Default no-op. */
-    virtual void setBufferWraps(const QStringList& wraps)
-    {
-        Q_UNUSED(wraps)
-    }
-    /** Buffer channel filter: "nearest", "linear", or "mipmap". Default no-op. */
-    virtual void setBufferFilter(const QString& filter)
-    {
-        Q_UNUSED(filter)
-    }
-    /** Per-channel buffer filters (up to 4). Empty = all use setBufferFilter value. Default no-op. */
-    virtual void setBufferFilters(const QStringList& filters)
-    {
-        Q_UNUSED(filters)
-    }
-
-    // Shader loading (paths; RHI node bakes GLSL 330 at runtime)
-    virtual bool loadVertexShader(const QString& path) = 0;
-    virtual bool loadFragmentShader(const QString& path) = 0;
-    virtual void setVertexShaderSource(const QString& source) = 0;
-    virtual void setFragmentShaderSource(const QString& source) = 0;
-
-    virtual bool isShaderReady() const = 0;
-    virtual QString shaderError() const = 0;
-    virtual void invalidateShader() = 0;
-    virtual void invalidateUniforms() = 0;
-};
-
-} // namespace PlasmaZones
